dh023

+++ src/main/java/ch/ethz/idsc/retina/dev/lidar/urg04lx/Urg04lxEvent.java
@@ -3,9 +3,9 @@ package ch.ethz.idsc.retina.dev.lidar.urg04lx;
 
 public class Urg04lxEvent {
   public final long timestamp;
-  public final short[] range;
+  public final double[] range;
 
-  public Urg04lxEvent(long timestamp, short[] range) {
+  public Urg04lxEvent(long timestamp, double[] range) {
     this.timestamp = timestamp;
     this.range = range;
   }
